move code to seahawk_rov
